+++ aten/src/ATen/native/Blas.cpp
@@ -10,6 +10,7 @@
 #include <ATen/native/mkldnn/Linear.h>
 #include <ATen/native/Resize.h>
 #include <ATen/native/GroupedMMUtils.h>
+#include <ATen/BlasBackend.h>
 #if !defined(__s390x__) && !defined(__powerpc__)
 #include <cpuinfo.h>
 #endif
